Separated template replace tests
